+++ cli/src/pcluster/models/cluster.py
@@ -440,6 +440,9 @@ class Cluster:
             Cluster._load_additional_instance_type_data(cluster_config_dict)
             config = self._load_config(cluster_config_dict)
             config.official_ami = self.__official_ami
+            if context.during_update:
+                config.managed_head_node_security_group = self.stack.get_resource_physical_id("HeadNodeSecurityGroup")
+                config.managed_compute_security_group = self.stack.get_resource_physical_id("ComputeSecurityGroup")
 
             validation_failures = config.validate(validator_suppressors, context)
             if any(f.level.value >= FailureLevel(validation_failure_level).value for f in validation_failures):
@@ -591,7 +594,8 @@ class Cluster:
                 # A nosec comment is appended to the following line in order to disable the B324 checks.
                 # The sha1 is used just as a hashing function.
                 # [B324:hashlib] Use of weak MD4, MD5, or SHA1 hash for security. Consider usedforsecurity=False
-                lambda value: hashlib.sha1(value.encode())  # nosec B324 nosemgrep
+                # [B303:blacklist] Use of insecure MD2, MD4, MD5, or SHA1 hash function
+                lambda value: hashlib.sha1(value.encode())  # nosec nosemgrep
                 .hexdigest()[0:16]
                 .capitalize()
             )
@@ -850,7 +854,7 @@ class Cluster:
             validator_suppressors=validator_suppressors,
             validation_failure_level=validation_failure_level,
             config_text=target_source_config,
-            context=ValidatorContext(head_node_instance_id=self.head_node_instance.id),
+            context=ValidatorContext(head_node_instance_id=self.head_node_instance.id, during_update=True),
         )
         changes = self._validate_patch(force, target_config)