+++ cli/src/pcluster/validators/awsbatch_validators.py
@@ -19,19 +19,6 @@ from pcluster.validators.common import FailureLevel, Validator
 LOGGER = logging.getLogger(__name__)
 
 
-class AwsBatchRegionValidator(Validator):
-    """
-    AWS Batch region validator.
-
-    Validate if the region is supported by AWS Batch.
-    """
-
-    def _validate(self, region: str):
-        # TODO use dryrun
-        if region in ["ap-northeast-3"]:
-            self._add_failure(f"AWS Batch scheduler is not supported in region '{region}'.", FailureLevel.ERROR)
-
-
 class AwsBatchComputeResourceSizeValidator(Validator):
     """
     AwsBatch compute resource size validator.