+++ cli/tests/pcluster/templates/test_capacity_reservation.py
@@ -24,9 +24,7 @@ from tests.pcluster.utils import get_head_node_policy, get_statement_by_sid
 
 @pytest.mark.parametrize(
     "config_file_name,",
-    [
-        ("config.yaml"),
-    ],
+    ["config.yaml"],
 )
 def test_capacity_reservation_id_permissions(mocker, test_datadir, config_file_name):
     mock_aws_api(mocker)
@@ -35,6 +33,7 @@ def test_capacity_reservation_id_permissions(mocker, test_datadir, config_file_n
         "pcluster.aws.ec2.Ec2Client.describe_capacity_reservations",
         return_value=[
             {
+                "CapacityReservationArn": "arn:partition:service:region:account-id:capacity-reservation/cr-12345",
                 "InstanceType": "c5.xlarge",
                 "AvailabilityZone": "us-east-1a",
             }