+++ tests/integration-tests/conftest.py
@@ -16,7 +16,6 @@
 import json
 import logging
 import os
-import random
 import re
 from functools import partial
 from itertools import product
@@ -28,8 +27,9 @@ from typing import Any, Optional, Tuple
 import boto3
 import pytest
 import yaml
+from _pytest._code import ExceptionInfo
 from _pytest.fixtures import FixtureDef, SubRequest
-from cfn_stacks_factory import CfnStack, CfnStacksFactory
+from cfn_stacks_factory import CfnStack, CfnStacksFactory, CfnVpcStack
 from clusters_factory import Cluster, ClustersFactory
 from conftest_markers import (
     DIMENSIONS_MARKER_ARGS,
@@ -49,8 +49,6 @@ from framework.tests_configuration.config_utils import get_all_regions
 from images_factory import Image, ImagesFactory
 from jinja2 import FileSystemLoader
 from jinja2.sandbox import SandboxedEnvironment
-from network_template_builder import Gateways, NetworkTemplateBuilder, SubnetConfig, VPCConfig
-from retrying import retry
 from troposphere import Ref, Sub, Template, ec2, resourcegroups
 from troposphere.ec2 import PlacementGroup
 from troposphere.efs import FileSystem as EFSFileSystem
@@ -66,6 +64,7 @@ from troposphere.fsx import (
 )
 from utils import (
     InstanceTypesData,
+    SetupError,
     create_s3_bucket,
     delete_s3_bucket,
     dict_add_nested_key,
@@ -80,6 +79,7 @@ from utils import (
     random_alphanumeric,
     scheduler_plugin_definition_uploader,
     set_logger_formatter,
+    to_pascal_case,
 )
 
 from tests.common.osu_common import run_osu_benchmarks
@@ -94,6 +94,8 @@ from tests.common.utils import (
     retrieve_pcluster_ami_without_standard_naming,
 )
 from tests.storage.snapshots_factory import EBSSnapshotsFactory
+
+pytest_plugins = ["conftest_networking"]
 
 
 def pytest_addoption(parser):
@@ -170,6 +172,10 @@ def pytest_addoption(parser):
     parser.addoption(
         "--slurm-database-stack-name",
         help="Name of CFN stack providing database stack to be used for testing Slurm accounting feature.",
+    )
+    parser.addoption(
+        "--external-shared-storage-stack-name",
+        help="Name of existing external shared storage stack.",
     )
 
 
@@ -757,9 +763,13 @@ def _add_policy_for_pre_post_install(node_config, custom_option, request, region
                 dict_add_nested_key(node_config, [additional_iam_policies], ("Iam", "AdditionalIamPolicies"))
 
 
-def _get_default_template_values(vpc_stack, request):
+def _get_default_template_values(vpc_stack: CfnVpcStack, request):
     """Build a dictionary of default values to inject in the jinja templated cluster configs."""
     default_values = get_vpc_snakecase_value(vpc_stack)
+    default_values["public_subnet_id"] = vpc_stack.get_public_subnet()  # TODO possibility to override default AZ
+    default_values["public_subnet_ids"] = vpc_stack.get_all_public_subnets()
+    default_values["private_subnet_id"] = vpc_stack.get_private_subnet()  # TODO possibility to override default AZ
+    default_values["private_subnet_ids"] = vpc_stack.get_all_private_subnets()
     default_values.update({dimension: request.node.funcargs.get(dimension) for dimension in DIMENSIONS_MARKER_ARGS})
     default_values["key_name"] = request.config.getoption("key_name")
 
@@ -815,57 +825,6 @@ def parameterized_cfn_stacks_factory(request):
     factory.delete_all_stacks()
 
 
-AVAILABILITY_ZONE_OVERRIDES = {
-    # c5.xlarge is not supported in use1-az3
-    # FSx Lustre file system creation is currently not supported for use1-az3
-    # p4d.24xlarge targeted ODCR is only available on use1-az6
-    "us-east-1": ["use1-az6"],
-    # some instance type is only supported in use2-az2
-    "us-east-2": ["use2-az2"],
-    # trn available on usw2-az4
-    "us-west-2": ["usw2-az4"],
-    # c5.xlarge is not supported in apse2-az3
-    "ap-southeast-2": ["apse2-az1", "apse2-az2"],
-    # FSx for Luster is not supported in apne1-az1
-    "ap-northeast-1": ["apne1-az4", "apne1-az2"],
-    # c4.xlarge is not supported in apne2-az2
-    "ap-northeast-2": ["apne2-az1", "apne2-az3"],
-    # c5.xlarge is not supported in apse1-az3
-    "ap-southeast-1": ["apse1-az2", "apse1-az1"],
-    # c4.xlarge is not supported in aps1-az2
-    "ap-south-1": ["aps1-az1", "aps1-az3"],
-    # NAT Gateway not available in sae1-az2 , c5n.18xlarge is not supported in sae1-az3
-    "sa-east-1": ["sae1-az1"],
-    # m6g.xlarge instances not available in euw1-az3
-    "eu-west-1": ["euw1-az1", "euw1-az2"],
-    # io2 EBS volumes not available in cac1-az4
-    "ca-central-1": ["cac1-az1", "cac1-az2"],
-    # instance can only be launch in placement group in eun1-az2
-    "eu-north-1": ["eun1-az2"],
-    # g3.8xlarge is not supported in euc1-az1
-    "eu-central-1": ["euc1-az2", "euc1-az3"],
-    # FSx not available in cnn1-az4
-    "cn-north-1": ["cnn1-az1", "cnn1-az2"],
-}
-
-
-@pytest.fixture(scope="function")
-def random_az_selector(request):
-    """Select random AZs for a given region."""
-
-    def _get_random_availability_zones(region, num_azs=1, default_value=None):
-        """Return num_azs random AZs (in the form of AZ names, e.g. 'us-east-1a') for the given region."""
-        az_ids = AVAILABILITY_ZONE_OVERRIDES.get(region, [])
-        if az_ids:
-            az_id_to_az_name_map = get_az_id_to_az_name_map(region, request.config.getoption("credential"))
-            sample = random.sample([az_id_to_az_name_map.get(az_id, default_value) for az_id in az_ids], k=num_azs)
-        else:
-            sample = [default_value] * num_azs
-        return sample[0] if num_azs == 1 else sample
-
-    return _get_random_availability_zones
-
-
 @pytest.fixture(scope="class", autouse=True)
 def setup_credentials(region, request):
     """Setup environment for the integ tests"""
@@ -881,40 +840,6 @@ def setup_env_variable(region):
     os.environ["AWS_DEFAULT_REGION"] = region
     yield
     del os.environ["AWS_DEFAULT_REGION"]
-
-
-def get_az_id_to_az_name_map(region, credential):
-    """Return a dict mapping AZ IDs (e.g, 'use1-az2') to AZ names (e.g., 'us-east-1c')."""
-    # credentials are managed manually rather than via setup_sts_credentials because this function
-    # is called by a session-scoped fixture, which cannot make use of a class-scoped fixture.
-    with aws_credential_provider(region, credential):
-        ec2_client = boto3.client("ec2", region_name=region)
-        return {
-            entry.get("ZoneId"): entry.get("ZoneName")
-            for entry in ec2_client.describe_availability_zones().get("AvailabilityZones")
-        }
-
-
-def get_availability_zones(region, credential):
-    """
-    Return a list of availability zones for the given region.
-
-    Note that this function is called by the vpc_stacks fixture. Because vcp_stacks is session-scoped,
-    it cannot utilize setup_sts_credentials, which is required in opt-in regions in order to call
-    describe_availability_zones.
-    """
-    az_list = []
-    with aws_credential_provider(region, credential):
-        client = boto3.client("ec2", region_name=region)
-        response_az = client.describe_availability_zones(
-            Filters=[
-                {"Name": "region-name", "Values": [str(region)]},
-                {"Name": "zone-type", "Values": ["availability-zone"]},
-            ]
-        )
-        for az in response_az.get("AvailabilityZones"):
-            az_list.append(az.get("ZoneName"))
-    return az_list
 
 
 @xdist_session_fixture(autouse=True)
@@ -964,126 +889,6 @@ def register_cli_credentials(initialize_cli_creds):
             register_cli_credentials_for_region(region, creds)
 
 
-@pytest.fixture(scope="session", autouse=True)
-def vpc_stacks(cfn_stacks_factory, request):
-    """Create VPC used by integ tests in all configured regions."""
-    regions = request.config.getoption("regions") or get_all_regions(request.config.getoption("tests_config"))
-    vpc_stacks = {}
-
-    for region in regions:
-        # Creating private_subnet_different_cidr in a different AZ for test_efs
-        # To-do: isolate this logic and create a compute subnet in different AZ than head node in test_efs
-
-        # if region has a non-empty list in AVAILABILITY_ZONE_OVERRIDES, select a subset of those AZs
-        credential = request.config.getoption("credential")
-        az_ids_for_region = AVAILABILITY_ZONE_OVERRIDES.get(region, [])
-        if az_ids_for_region:
-            az_id_to_az_name = get_az_id_to_az_name_map(region, credential)
-            az_names = [az_id_to_az_name.get(az_id) for az_id in az_ids_for_region]
-            # if only one AZ can be used for the given region, use it multiple times
-            if len(az_names) == 1:
-                availability_zones = az_names * 3
-            if len(az_names) == 2:
-                # ensures that az[0] and az[1] are always different if two az are available for use
-                availability_zones = az_names + random.sample(az_names, k=1)
-        # otherwise, select a subset of all AZs in the region
-        else:
-            az_list = get_availability_zones(region, credential)
-            # if number of available zones is smaller than 3, list is expanded to 3 and filled with [None, ...]
-            if len(az_list) < 3:
-                diff = 3 - len(az_list)
-                availability_zones = az_list + [None] * diff
-            else:
-                availability_zones = random.sample(az_list, k=3)
-
-        # Subnets visual representation:
-        # http://www.davidc.net/sites/default/subnets/subnets.html?network=192.168.0.0&mask=16&division=7.70
-        public_subnet = SubnetConfig(
-            name="Public",
-            cidr="192.168.32.0/20",  # 4096 IPs
-            map_public_ip_on_launch=True,
-            has_nat_gateway=True,
-            availability_zone=availability_zones[0],
-            default_gateway=Gateways.INTERNET_GATEWAY,
-        )
-        private_subnet = SubnetConfig(
-            name="Private",
-            cidr="192.168.64.0/20",  # 4096 IPs
-            map_public_ip_on_launch=False,
-            has_nat_gateway=False,
-            availability_zone=availability_zones[0],
-            default_gateway=Gateways.NAT_GATEWAY,
-        )
-        private_subnet_different_cidr = SubnetConfig(
-            name="PrivateAdditionalCidr",
-            cidr="192.168.96.0/20",  # 4096 IPs
-            map_public_ip_on_launch=False,
-            has_nat_gateway=False,
-            availability_zone=availability_zones[1],
-            default_gateway=Gateways.NAT_GATEWAY,
-        )
-        no_internet_subnet = SubnetConfig(
-            name="NoInternet",
-            cidr="192.168.16.0/20",  # 4096 IPs
-            map_public_ip_on_launch=False,
-            has_nat_gateway=False,
-            availability_zone=availability_zones[0],
-            default_gateway=Gateways.NONE,
-        )
-        public_subnet_az2 = SubnetConfig(
-            name="PublicAz2",
-            cidr="192.168.128.0/20",  # 4096 IPs
-            map_public_ip_on_launch=True,
-            has_nat_gateway=True,
-            availability_zone=availability_zones[1],
-            default_gateway=Gateways.INTERNET_GATEWAY,
-        )
-        private_subnet_az2 = SubnetConfig(
-            name="PrivateAz2",
-            cidr="192.168.160.0/20",  # 4096 IPs
-            map_public_ip_on_launch=False,
-            has_nat_gateway=False,
-            availability_zone=availability_zones[1],
-            default_gateway=Gateways.NAT_GATEWAY,
-        )
-        public_subnet_az3 = SubnetConfig(
-            name="PublicAz3",
-            cidr="192.168.192.0/20",  # 4096 IPs
-            map_public_ip_on_launch=True,
-            has_nat_gateway=True,
-            availability_zone=availability_zones[2],
-            default_gateway=Gateways.INTERNET_GATEWAY,
-        )
-        private_subnet_az3 = SubnetConfig(
-            name="PrivateAz3",
-            cidr="192.168.224.0/20",  # 4096 IPs
-            map_public_ip_on_launch=False,
-            has_nat_gateway=False,
-            availability_zone=availability_zones[2],
-            default_gateway=Gateways.NAT_GATEWAY,
-        )
-        vpc_config = VPCConfig(
-            cidr="192.168.0.0/17",
-            additional_cidr_blocks=["192.168.128.0/17"],
-            subnets=[
-                public_subnet,
-                private_subnet,
-                private_subnet_different_cidr,
-                no_internet_subnet,
-                public_subnet_az2,
-                private_subnet_az2,
-                public_subnet_az3,
-                private_subnet_az3,
-            ],
-        )
-        template = NetworkTemplateBuilder(
-            vpc_configuration=vpc_config, default_availability_zone=availability_zones[0]
-        ).build()
-        vpc_stacks[region] = _create_vpc_stack(request, template, region, cfn_stacks_factory)
-
-    return vpc_stacks
-
-
 @pytest.fixture(scope="class")
 @pytest.mark.usefixtures("clusters_factory", "images_factory")
 def create_roles_stack(request, region):
@@ -1115,11 +920,6 @@ def create_roles_stack(request, region):
         logging.warning("Skipping deletion of IAM roles stack because --no-delete option is set")
 
 
-@pytest.fixture(scope="class")
-def vpc_stack(vpc_stacks, region):
-    return vpc_stacks[region]
-
-
 @pytest.fixture(scope="session")
 def public_ecr_image_uri(request):
     return request.config.getoption("public_ecr_image_uri")
@@ -1138,27 +938,6 @@ def api_definition_s3_uri(request):
 @pytest.fixture(scope="session")
 def api_infrastructure_s3_uri(request):
     return request.config.getoption("api_infrastructure_s3_uri")
-
-
-# If stack creation fails it'll retry once more. This is done to mitigate failures due to resources
-# not available in randomly picked AZs.
-@retry(
-    stop_max_attempt_number=2,
-    wait_fixed=5000,
-    retry_on_exception=lambda exception: not isinstance(exception, KeyboardInterrupt),
-)
-def _create_vpc_stack(request, template, region, cfn_stacks_factory):
-    if request.config.getoption("vpc_stack"):
-        logging.info("Using stack {0} in region {1}".format(request.config.getoption("vpc_stack"), region))
-        stack = CfnStack(name=request.config.getoption("vpc_stack"), region=region, template=template.to_json())
-    else:
-        stack = CfnStack(
-            name=generate_stack_name("integ-tests-vpc", request.config.getoption("stackname_suffix")),
-            region=region,
-            template=template.to_json(),
-        )
-        cfn_stacks_factory.create_stack(stack)
-    return stack
 
 
 @pytest.fixture(scope="class")
@@ -1281,6 +1060,9 @@ def pytest_runtest_makereport(item, call):
     setattr(item, "rep_" + rep.when, rep)
 
     if rep.when in ["setup", "call"] and rep.failed:
+        exception_info: ExceptionInfo = call.excinfo
+        if exception_info.value and isinstance(exception_info.value, SetupError):
+            rep.when = "setup"
         try:
             update_failed_tests_config(item)
         except Exception as e:
@@ -1383,17 +1165,14 @@ def placement_group_stack(cfn_stacks_factory, request, region):
 
 
 @pytest.fixture(scope="class")
-def odcr_stack(request, region, placement_group_stack, cfn_stacks_factory, vpc_stack):
+def odcr_stack(request, region, placement_group_stack, cfn_stacks_factory, vpc_stack: CfnVpcStack):
     logging.info("Setting up the ODCR stack")
     odcr_template = Template()
     odcr_template.set_version()
     odcr_template.set_description("ODCR stack to test open, targeted, and PG ODCRs")
-    availability_zone = (
-        boto3.resource("ec2").Subnet(get_vpc_snakecase_value(vpc_stack)["public_subnet_id"]).availability_zone
-    )
-    availability_zone_2 = (
-        boto3.resource("ec2").Subnet(get_vpc_snakecase_value(vpc_stack)["public_az2_subnet_id"]).availability_zone
-    )
+    public_subnets_list = vpc_stack.get_all_public_subnets()
+    availability_zone = boto3.resource("ec2").Subnet(public_subnets_list[0]).availability_zone
+    availability_zone_2 = boto3.resource("ec2").Subnet(public_subnets_list[1]).availability_zone
     open_odcr = ec2.CapacityReservation(
         "integTestsOpenOdcr",
         AvailabilityZone=availability_zone,
@@ -1627,12 +1406,6 @@ def mpi_variants(architecture):
     return variants
 
 
-def to_pascal_case(snake_case_word):
-    """Convert the given snake case word into a PascalCase one."""
-    parts = iter(snake_case_word.split("_"))
-    return "".join(word.title() for word in parts)
-
-
 @pytest.fixture()
 def run_benchmarks(request, mpi_variants, test_datadir, instance, os, region, benchmarks):
     def _run_benchmarks(remote_command_executor, scheduler_commands, **kwargs):
@@ -1731,7 +1504,7 @@ def scheduler_commands_factory(scheduler, scheduler_plugin_configuration):
 
 
 @pytest.fixture(scope="class")
-def fsx_factory(vpc_stack, cfn_stacks_factory, request, region, key_name):
+def fsx_factory(vpc_stack: CfnVpcStack, cfn_stacks_factory, request, region, key_name):
     """
     Define a fixture to manage the creation and destruction of fsx.
 
@@ -1774,7 +1547,7 @@ def fsx_factory(vpc_stack, cfn_stacks_factory, request, region, key_name):
             fsx_filesystem = FileSystem(
                 title=f"{file_system_resource_name}{i}",
                 SecurityGroupIds=[Ref(fsx_sg)],
-                SubnetIds=[vpc_stack.cfn_outputs["PublicSubnetId"]],
+                SubnetIds=[vpc_stack.get_public_subnet()],
                 FileSystemType=file_system_type,
                 **kwargs,
                 **depends_on_arg,
@@ -1975,7 +1748,7 @@ def efs_mount_target_stack_factory(cfn_stacks_factory, request, region, vpc_stac
             )
 
         # Create mount targets
-        subnet_ids = [value for key, value in vpc_stack.cfn_outputs.items() if key.endswith("SubnetId")]
+        subnet_ids = vpc_stack.get_all_public_subnets() + vpc_stack.get_all_private_subnets()
         _add_mount_targets(subnet_ids, efs_ids, security_group, template)
 
         stack_name = generate_stack_name("integ-tests-mount-targets", request.config.getoption("stackname_suffix"))